refactor(solver): split partially inferable inference helper
